Change shared folder to be shared from guest to host.
It turns out the fact that Windows is using ntfs as the underlying file
system was breaking pretty much everything with npm / other things that
do things with file permissions that don't work the same way in ntfs.

To fix this, changes were made to have the guest OS host a samba server
to be used as the shared folder, and the documentation was changed
accordingly.

Also, since we're not supporting embedded projects anymore, remove a lot
of the packages added for hibike / smartsensors / etc.